drm/amd/dal: fix to enable correct i2c hw buffer
In this change I2C HW buffer size changes to 538 (for >= DCE11)

Also, changed usage of uint8_t length to uint32_t to
enable i2c reads of more then 256 bytes

Acked-by: Harry Wentland <harry.wentland@amd.com>

+++ drivers/gpu/drm/amd/dal/dc/dc_ddc_types.h
@@ -28,7 +28,7 @@
 struct i2c_payload {
 	bool write;
 	uint8_t address;
-	uint8_t length;
+	uint32_t length;
 	uint8_t *data;
 };
